+++ packages/components-js/tests/e2e/specs/link-tile.e2e.ts
@@ -1,5 +1,5 @@
-import type { Page } from 'playwright';
 import { expect, test } from '@playwright/test';
+import type { Page } from 'playwright';
 import { getLifecycleStatus, setContentWithDesignSystem, setProperty, waitForStencilLifecycle } from '../helpers';
 
 const getHost = (page: Page) => page.locator('p-link-tile');